0224 2교시

+++ src/main/java/com/project/reservation/repository/ReviewRepository.java
@@ -33,5 +33,5 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
 
     // 리뷰의 총 좋아요 수 조회
     @Query(value = "SELECT COUNT(l) FROM ReviewLike l WHERE l.review.id = :reviewId")
-    long countLikesByReview(Long reviewId);
+    int countLikesByReview(Long reviewId);
 }
